Added save_arrhenius and import_arrhenius functions + created analysis notebook for arrhenius comparison
These functions allow to save and then read the arrhenius info such as the actual arrhenius plot data and the trap parameters (Ea, sigma and errors).
This will allow to simplify further analysis of the arrhenius plots in order to compare different samples.

I also started to create the notebook that
compares arrhenius plots of all samples, with statistical analysis of the data.