Removed default line normalization + drop option in plot_arrhenius
Now the lines in the picts map are by default non-normalized to 1. We found out that normalizing to one the lines led to artefacts in the map. We introduced the line_normalize parameter, which, if set to True, allows to calculate the normalized map.

Sometimes one does not want to display in the arrhenius plot all the traps, so we introduced the "drop" parameter in plot_arrhenius, which takes a list of the trap names that you don't want to be displayed in the plot.